move compareDates script to its own file
